+++ ChangeLog.md
@@ -88,5 +88,10 @@ Testade att fetcha länder från API till database tabellen. Det blev en hel del
 
 # 17/12
 
+- GET-POST-DELETE TravelList
 - Komponenter Karta, CountryModal, TravelList
 - Komponent test TravelList
+
+# 18/12
+
+- Komplett E2E test för TravelList-DELETE